Small fixes/changes, Added extra config feature : keep chunks with active furnaces online.

+++ src/com/guntherdw/bukkit/tweakcraft/Util/TimeTool.java
@@ -23,7 +23,7 @@ package com.guntherdw.bukkit.tweakcraft.Util;
  */
 public class TimeTool {
     public static String calcLeft(Long seconds) {
-        if(seconds > Integer.MAX_VALUE)
+        if(seconds > Long.MAX_VALUE)
             return null;
 
         Integer days = 0;
